Make the remaining int narrowing explicit
The value conversion class stored int results in byte, ubyte, char,
sh_int, and sbyte fields: race and weapon tables, clan privileges,
preferences, conditions, room light, affect locations. The fields that
hold running totals were widened in step 1.2, so these take explicit
casts that keep the current truncation.

380 sites are cast from the Clang excerpts with the step 1.3 range
script; tolower and toupper results, RANGE, LIMIT, and GET_LEVEL
assignments, the summon damage dice, the affect location and modifier
arguments, map coordinates, and htons ports are cast by hand.
MOB_SET_FEAT, SET_ABILITY, and VESSEL_REPAIR_FIELD cast inside the macro.
GCC's compound assignments into narrow fields read x = (T)(x op (y)).

dc_bonus is an int now: 47 compound updates adjust the spell DC bonus
in it, and as a byte it wrapped past 127. The logon record keeps its int
fields because struct last_entry is written to disk with fwrite.

check_warning_budget.py accepts a log without warnings when the budget
file lists no classes and the log shows compilation, so a clean
migration build passes the budget job; while classes remain, an empty
log still fails. Both budget files are now empty.

+++ src/character/feats.c
@@ -326,9 +326,9 @@ static void feato(int featnum, const char *name, int in_game, int can_learn, int
                   int feat_type, const char *short_description, const char *description)
 {
   feat_list[featnum].name = name;
-  feat_list[featnum].in_game = in_game;
-  feat_list[featnum].can_learn = can_learn;
-  feat_list[featnum].can_stack = can_stack;
+  feat_list[featnum].in_game = (sbyte)in_game;
+  feat_list[featnum].can_learn = (sbyte)can_learn;
+  feat_list[featnum].can_stack = (sbyte)can_stack;
   feat_list[featnum].feat_type = feat_type;
   feat_list[featnum].short_description = short_description;
   feat_list[featnum].description = description;
